+++ test/test_debug_multifile.ml
@@ -4,7 +4,7 @@ module Debug_runtime =
   (* Split as soon as possible. *)
   (val Minidebug_runtime.debug_file ~hyperlink:"../" ~split_files_after:(1 lsl 9)
          ~for_append:false ~backend:(`Markdown Minidebug_runtime.default_md_config)
-         ~values_first_mode:true "debugger_multifile")
+         "debugger_multifile")
 
 let%debug_sexp foo (x : int) : int list =
   let y : int = x + 1 in