+++ test/test_expect_test.ml
@@ -191,7 +191,7 @@ let%expect_test "%debug_show flushing with global prefix" =
     |}]
 
 let%expect_test "%debug_show disabled subtree" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%debug_show rec loop_complete (x : int) : int =
     let z : int = (x - 1) / 2 in
     if x <= 0 then 0 else z + loop_complete (z + (x / 2))
@@ -271,7 +271,7 @@ let%expect_test "%debug_show disabled subtree" =
     |}]
 
 let%expect_test "%debug_show with exception" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%debug_show rec loop_truncated (x : int) : int =
     let z : int = (x - 1) / 2 in
     if x <= 0 then failwith "the log as for loop_complete but without return values";
@@ -320,7 +320,7 @@ let%expect_test "%debug_show with exception" =
     |}]
 
 let%expect_test "%debug_show depth exceeded" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%debug_show rec loop_exceeded (x : int) : int =
     [%debug_interrupts
       { max_nesting_depth = 5; max_num_children = 1000 };
@@ -358,7 +358,7 @@ let%expect_test "%debug_show depth exceeded" =
     |}]
 
 let%expect_test "%debug_show num children exceeded linear" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let () =
     try
       let%debug_show _bar : unit =
@@ -403,7 +403,7 @@ let%expect_test "%debug_show num children exceeded linear" =
     |}]
 
 let%expect_test "%debug_show truncated children linear" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~truncate_children:10 ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ~truncate_children:10 ()) in
   let () =
     try
       let%debug_show _bar : unit =
@@ -442,7 +442,7 @@ let%expect_test "%debug_show truncated children linear" =
     |}]
 
 let%expect_test "%track_show track for-loop num children exceeded" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let () =
     try
       let%track_show _bar : unit =
@@ -487,7 +487,7 @@ let%expect_test "%track_show track for-loop num children exceeded" =
     |}]
 
 let%expect_test "%track_show track for-loop truncated children" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~truncate_children:10 ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ~truncate_children:10 ()) in
   let () =
     try
       let%track_show _bar : unit =
@@ -529,7 +529,7 @@ let%expect_test "%track_show track for-loop truncated children" =
     |}]
 
 let%expect_test "%track_show track for-loop" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let () =
     try
       let%track_show _bar : unit =
@@ -580,7 +580,7 @@ let%expect_test "%track_show track for-loop" =
     |}]
 
 let%expect_test "%track_show track for-loop, time spans" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~elapsed_times:Microseconds ())
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ~elapsed_times:Microseconds ())
   in
   let () =
     try
@@ -637,7 +637,7 @@ let%expect_test "%track_show track for-loop, time spans" =
     |}]
 
 let%expect_test "%track_show track while-loop" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let () =
     try
       let%track_show _bar : unit =
@@ -677,7 +677,7 @@ let%expect_test "%track_show track while-loop" =
     |}]
 
 let%expect_test "%debug_show num children exceeded nested" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%debug_show rec loop_exceeded (x : int) : int =
     [%debug_interrupts
       { max_nesting_depth = 1000; max_num_children = 10 };
@@ -734,7 +734,7 @@ let%expect_test "%debug_show num children exceeded nested" =
     |}]
 
 let%expect_test "%debug_show truncated children nested" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~truncate_children:4 ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ~truncate_children:4 ()) in
   let%debug_show rec loop_exceeded (x : int) : int =
     Array.fold_left ( + ) 0
     @@ Array.init
@@ -862,7 +862,7 @@ let%expect_test "%debug_show truncated children nested" =
 
 let%expect_test "%track_show highlight" =
   let module Debug_runtime =
-    (val Minidebug_runtime.debug ~highlight_terms:(Re.str "3") ())
+    (val Minidebug_runtime.debug ~values_first_mode:false ~highlight_terms:(Re.str "3") ())
   in
   let%debug_show rec loop_highlight (x : int) : int =
     [%debug_interrupts
@@ -934,7 +934,7 @@ let%expect_test "%track_show highlight" =
     |}]
 
 let%expect_test "%track_show PrintBox tracking" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%track_show track_branches (x : int) : int =
     if x < 6 then match x with 0 -> 1 | 1 -> 0 | _ -> ~-x
     else match x with 6 -> 5 | 7 -> 4 | _ -> x
@@ -963,7 +963,7 @@ let%expect_test "%track_show PrintBox tracking" =
     |}]
 
 let%expect_test "%track_show PrintBox tracking <function>" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%track_show track_branches = function
     | 0 -> 1
     | 1 -> 0
@@ -988,7 +988,7 @@ let%expect_test "%track_show PrintBox tracking <function>" =
     |}]
 
 let%expect_test "%track_show PrintBox tracking with debug_notrace" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%track_show track_branches (x : int) : int =
     if x < 6 then
       match%debug_notrace x with
@@ -1034,7 +1034,7 @@ let%expect_test "%track_show PrintBox tracking with debug_notrace" =
 
 let%expect_test "%track_show PrintBox not tracking anonymous functions with debug_notrace"
     =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%track_show track_foo (x : int) : int =
     [%debug_notrace (fun (y : int) -> ignore y) x];
     let w = [%debug_notrace (fun (v : int) -> v) x] in
@@ -1057,7 +1057,7 @@ let%expect_test "%track_show PrintBox not tracking anonymous functions with debu
     |}]
 
 let%expect_test "respect scope of nested extension points" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%track_show track_branches (x : int) : int =
     if x < 6 then
       match%debug_notrace x with
@@ -1101,7 +1101,7 @@ let%expect_test "respect scope of nested extension points" =
     |}]
 
 let%expect_test "%debug_show un-annotated toplevel fun" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%debug_show anonymous x =
     let nested y = y + 1 in
     [%log "We do log this function"];
@@ -1126,7 +1126,7 @@ let%expect_test "%debug_show un-annotated toplevel fun" =
     |}]
 
 let%expect_test "%debug_show nested un-annotated toplevel fun" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%debug_show wrapper () =
     let%debug_show anonymous x =
       let nested y = y + 1 in
@@ -1155,7 +1155,7 @@ let%expect_test "%debug_show nested un-annotated toplevel fun" =
     |}]
 
 let%expect_test "%track_show no return type anonymous fun" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%debug_show anonymous (x : int) =
     Array.fold_left ( + ) 0 @@ Array.init (x + 1) (fun (i : int) -> i)
   in
@@ -1193,7 +1193,7 @@ let%expect_test "%track_show no return type anonymous fun" =
     |}]
 
 let%expect_test "%track_show anonymous fun, num children exceeded" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%track_show rec loop_exceeded (x : int) : int =
     [%debug_interrupts
       { max_nesting_depth = 1000; max_num_children = 10 };
@@ -1289,7 +1289,7 @@ let%expect_test "%track_show anonymous fun, num children exceeded" =
     |}]
 
 let%expect_test "%track_show anonymous fun, truncated children" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~truncate_children:2 ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ~truncate_children:2 ()) in
   let%track_show rec loop_exceeded (x : int) : int =
     Array.fold_left ( + ) 0
     @@ Array.init
@@ -1347,7 +1347,7 @@ module type T = sig
 end
 
 let%expect_test "%debug_show function with abstract type" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%debug_show foo (type d) (module D : T with type c = d) ~a (c : int) : int =
     if c = 0 then 0 else List.length [ a; D.c ]
   in
@@ -1373,7 +1373,7 @@ let%expect_test "%debug_show function with abstract type" =
     |}]
 
 let%expect_test "%debug_show PrintBox values_first_mode to stdout with exception" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%debug_show rec loop_truncated (x : int) : int =
     let z : int = (x - 1) / 2 in
     if x <= 0 then failwith "the log as for loop_complete but without return values";
@@ -1431,7 +1431,7 @@ let%expect_test "%debug_show PrintBox values_first_mode to stdout with exception
 
 let%expect_test "%debug_show PrintBox values_first_mode to stdout num children exceeded \
                  linear" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let () =
     try
       let%debug_show _bar : unit =
@@ -1477,7 +1477,7 @@ let%expect_test "%debug_show PrintBox values_first_mode to stdout num children e
     |}]
 
 let%expect_test "%track_show PrintBox values_first_mode to stdout track for-loop" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let () =
     try
       let%track_show _bar : unit =
@@ -1537,7 +1537,7 @@ let%expect_test "%track_show PrintBox values_first_mode to stdout track for-loop
 
 let%expect_test "%debug_show PrintBox values_first_mode to stdout num children exceeded \
                  nested" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%debug_show rec loop_exceeded (x : int) : int =
     [%debug_interrupts
       { max_nesting_depth = 1000; max_num_children = 10 };
@@ -1600,7 +1600,7 @@ let%expect_test "%debug_show PrintBox values_first_mode to stdout num children e
 let%expect_test "%debug_show elapsed times PrintBox values_first_mode to stdout nested, \
                  truncated children" =
   let module Debug_runtime =
-    (val Minidebug_runtime.debug ~elapsed_times:Microseconds ~values_first_mode:true
+    (val Minidebug_runtime.debug ~elapsed_times:Microseconds
            ~truncate_children:4 ())
   in
   let%debug_show rec loop_exceeded (x : int) : int =
@@ -1765,7 +1765,7 @@ let%expect_test "%debug_show elapsed times PrintBox values_first_mode to stdout 
 
 let%expect_test "%debug_show PrintBox values_first_mode to stdout highlight" =
   let module Debug_runtime =
-    (val Minidebug_runtime.debug ~highlight_terms:(Re.str "3") ~values_first_mode:true ())
+    (val Minidebug_runtime.debug ~highlight_terms:(Re.str "3") ())
   in
   let%debug_show rec loop_highlight (x : int) : int =
     let z : int = (x - 1) / 2 in
@@ -1833,7 +1833,7 @@ let%expect_test "%debug_show PrintBox values_first_mode to stdout highlight" =
     |}]
 
 let%expect_test "%track_show PrintBox values_first_mode tracking" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%track_show track_branches (x : int) : int =
     if x < 6 then match x with 0 -> 1 | 1 -> 0 | _ -> ~-x
     else match x with 6 -> 5 | 7 -> 4 | _ -> x
@@ -1867,7 +1867,7 @@ let%expect_test "%track_show PrintBox values_first_mode tracking" =
 
 let%expect_test "%track_show PrintBox values_first_mode to stdout no return type \
                  anonymous fun" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%track_show anonymous (x : int) =
     Array.fold_left ( + ) 0 @@ Array.init (x + 1) (fun (i : int) -> i)
   in
@@ -1975,7 +1975,7 @@ let%expect_test "%debug_show tuples" =
     |}]
 
 let%expect_test "%debug_show records values_first_mode" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%debug_show bar { first : int; second : int } : int =
     let { first : int = a; second : int = b } = { first; second = second + 3 } in
     let y : int = a + 1 in
@@ -2015,7 +2015,7 @@ let%expect_test "%debug_show records values_first_mode" =
     |}]
 
 let%expect_test "%debug_show tuples values_first_mode" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%debug_show bar ((first : int), (second : int)) : int =
     let y : int = first + 1 in
     second * y
@@ -2067,7 +2067,7 @@ type ('a, 'b) left_right = Left of 'a | Right of 'b
 type ('a, 'b, 'c) one_two_three = One of 'a | Two of 'b | Three of 'c
 
 let%expect_test "%track_show variants values_first_mode" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%track_show bar (Zero (x : int)) : int =
     let y = (x + 1 : int) in
     2 * y
@@ -2111,7 +2111,7 @@ let%expect_test "%track_show variants values_first_mode" =
     |}]
 
 let%expect_test "%debug_show tuples merge type info" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%debug_show baz (((first : int), (second : 'a)) : 'b * int) : int * int =
     let ((y : 'c), (z : int)) : int * 'd = (first + 1, 3) in
     let (a : int), b = (first + 1, (second + 3 : int)) in
@@ -2145,7 +2145,7 @@ let%expect_test "%debug_show tuples merge type info" =
     |}]
 
 let%expect_test "%debug_show decompose multi-argument function type" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%debug_show f : 'a. 'a -> int -> int = fun _a b -> b + 1 in
   let%debug_show g : 'a. 'a -> int -> 'a -> 'a -> int = fun _a b _c _d -> b * 2 in
   let () = print_endline @@ Int.to_string @@ f 'a' 6 in
@@ -2164,7 +2164,7 @@ let%expect_test "%debug_show decompose multi-argument function type" =
     |}]
 
 let%expect_test "%debug_show debug type info" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   [%debug_show
     [%debug_type_info
       let f : 'a. 'a -> int -> int = fun _a b -> b + 1 in
@@ -2187,7 +2187,7 @@ let%expect_test "%debug_show debug type info" =
     |}]
 
 let%expect_test "%track_show options values_first_mode" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%track_show foo l : int =
     match (l : int option) with None -> 7 | Some y -> y * 2
   in
@@ -2232,7 +2232,7 @@ let%expect_test "%track_show options values_first_mode" =
     |}]
 
 let%expect_test "%track_show list values_first_mode" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%track_show foo l : int = match (l : int list) with [] -> 7 | y :: _ -> y * 2 in
   let () = print_endline @@ Int.to_string @@ foo [ 7 ] in
   let%track_show bar (l : int list) : int = match l with [] -> 7 | y :: _ -> y * 2 in
@@ -2286,7 +2286,7 @@ let%expect_test "%track_rt_show list runtime passing" =
     print_endline @@ Int.to_string
     @@ foo
          Minidebug_runtime.(
-           forget_printbox @@ debug ~global_prefix:"foo-1" ~values_first_mode:true ())
+           forget_printbox @@ debug ~global_prefix:"foo-1" ())
          [ 7 ]
   in
   let%track_rt_show baz : int list -> int = function
@@ -2299,14 +2299,14 @@ let%expect_test "%track_rt_show list runtime passing" =
     print_endline @@ Int.to_string
     @@ baz
          Minidebug_runtime.(
-           forget_printbox @@ debug ~global_prefix:"baz-1" ~values_first_mode:true ())
+           forget_printbox @@ debug ~global_prefix:"baz-1" ())
          [ 4 ]
   in
   let () =
     print_endline @@ Int.to_string
     @@ baz
          Minidebug_runtime.(
-           forget_printbox @@ debug ~global_prefix:"baz-2" ~values_first_mode:true ())
+           forget_printbox @@ debug ~global_prefix:"baz-2" ())
          [ 4; 5; 6 ]
   in
   [%expect
@@ -2369,7 +2369,7 @@ let%expect_test "%track_rt_show procedure runtime passing" =
     |}]
 
 let%expect_test "%track_rt_show nested procedure runtime passing" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%debug_show rt_test () =
     let%track_rt_show bar () = (fun () -> ()) () in
     let%track_rt_show foo () =
@@ -2409,7 +2409,7 @@ let%expect_test "%track_rt_show nested procedure runtime passing" =
     |}]
 
 let%expect_test "%log constant entries" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%debug_show foo () : unit =
     [%log "This is the first log line"];
     [%log [ "This is the"; "2"; "log line" ]];
@@ -2445,7 +2445,7 @@ let%expect_test "%log constant entries" =
     |}]
 
 let%expect_test "%log with type annotations" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let i = 3 in
   let pi = 3.14 in
   let l = [ 1; 2; 3 ] in
@@ -2470,7 +2470,7 @@ let%expect_test "%log with type annotations" =
     |}]
 
 let%expect_test "%log with default type assumption" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let s = "3" in
   let pi = "3.14" in
   let x2 s = "2*" ^ s in
@@ -2497,7 +2497,7 @@ let%expect_test "%log with default type assumption" =
     |}]
 
 let%expect_test "%log track while-loop" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%track_sexp result =
     let i = ref 0 in
     let j = ref 0 in
@@ -2560,28 +2560,28 @@ let%expect_test "%log runtime log levels while-loop" =
   print_endline
   @@ Int.to_string
        (result
-          Minidebug_runtime.(forget_printbox @@ debug ~global_prefix:"Everything" ())
+          Minidebug_runtime.(forget_printbox @@ debug ~values_first_mode:false ~global_prefix:"Everything" ())
           ());
   print_endline
   @@ Int.to_string
        (result
           Minidebug_runtime.(
             forget_printbox
-            @@ debug ~values_first_mode:true ~log_level:0 ~global_prefix:"Nothing" ())
+            @@ debug ~values_first_mode:false ~log_level:0 ~global_prefix:"Nothing" ())
           ());
   print_endline
   @@ Int.to_string
        (result
           Minidebug_runtime.(
             forget_printbox
-            @@ debug ~values_first_mode:true ~log_level:1 ~global_prefix:"Error" ())
+            @@ debug ~log_level:1 ~global_prefix:"Error" ())
           ());
   print_endline
   @@ Int.to_string
        (result
           Minidebug_runtime.(
             forget_printbox
-            @@ debug ~values_first_mode:true ~log_level:2 ~global_prefix:"Warning" ())
+            @@ debug ~log_level:2 ~global_prefix:"Warning" ())
           ());
   [%expect
     {|
@@ -2724,7 +2724,7 @@ let%expect_test "%log runtime log levels while-loop" =
     |}]
 
 let%expect_test "%log compile time log levels while-loop" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%track_sexp everything () : int =
     [%log_level
       9;
@@ -2866,7 +2866,7 @@ let%expect_test "%log compile time log levels while-loop" =
 
 let%expect_test "%log compile time log levels runtime-passing while-loop" =
   let module Debug_runtime =
-    (val Minidebug_runtime.debug ~global_prefix:"TOPLEVEL" ~values_first_mode:true ())
+    (val Minidebug_runtime.debug ~global_prefix:"TOPLEVEL" ())
   in
   (* Compile-time log level restrictions cannot be undone, since the logging code is not
      generated. *)
@@ -2889,7 +2889,7 @@ let%expect_test "%log compile time log levels runtime-passing while-loop" =
        print_endline @@ Int.to_string
        @@ nothing
             Minidebug_runtime.(
-              forget_printbox @@ debug ~global_prefix:"nothing" ~values_first_mode:true ())
+              forget_printbox @@ debug ~global_prefix:"nothing" ())
             ()]);
     [%log_level
       2;
@@ -2911,7 +2911,7 @@ let%expect_test "%log compile time log levels runtime-passing while-loop" =
       print_endline @@ Int.to_string
       @@ warning
            Minidebug_runtime.(
-             forget_printbox @@ debug ~global_prefix:"warning" ~values_first_mode:true ())
+             forget_printbox @@ debug ~global_prefix:"warning" ())
            ()]
   in
   [%expect
@@ -2952,7 +2952,7 @@ let%expect_test "%log compile time log levels runtime-passing while-loop" =
     |}]
 
 let%expect_test "%log track while-loop result" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%track_sexp result =
     let i = ref 0 in
     let j = ref 0 in
@@ -3008,7 +3008,7 @@ let%expect_test "%log track while-loop result" =
     |}]
 
 let%expect_test "%log without scope" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~print_entry_ids:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ~print_entry_ids:true ()) in
   let i = 3 in
   let pi = 3.14 in
   let l = [ 1; 2; 3 ] in
@@ -3045,7 +3045,7 @@ let%expect_test "%log without scope" =
 
 let%expect_test "%log without scope values_first_mode" =
   let module Debug_runtime =
-    (val Minidebug_runtime.debug ~print_entry_ids:true ~values_first_mode:true ())
+    (val Minidebug_runtime.debug ~print_entry_ids:true ())
   in
   let i = 3 in
   let pi = 3.14 in
@@ -3080,7 +3080,7 @@ let%expect_test "%log without scope values_first_mode" =
 
 let%expect_test "%log with print_entry_ids, mixed up scopes" =
   let module Debug_runtime =
-    (val Minidebug_runtime.debug ~print_entry_ids:true ~values_first_mode:true ())
+    (val Minidebug_runtime.debug ~print_entry_ids:true ())
   in
   let i = 3 in
   let pi = 3.14 in
@@ -3140,7 +3140,7 @@ let%expect_test "%log with print_entry_ids, verbose_entry_ids in HTML, values_fi
     =
   let module Debug_runtime =
     (val Minidebug_runtime.debug ~print_entry_ids:true ~verbose_entry_ids:true
-           ~values_first_mode:true ())
+           ())
   in
   Debug_runtime.config.backend <- `Html PrintBox_html.Config.default;
   let i = 3 in
@@ -3191,7 +3191,7 @@ let%expect_test "%log with print_entry_ids, verbose_entry_ids in HTML, values_fi
     |}]
 
 let%expect_test "%diagn_show ignores type annots" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%diagn_show toplevel =
     let bar { first : int; second : int } : int =
       let { first : int = a; second : int = b } = { first; second = second + 3 } in
@@ -3220,7 +3220,7 @@ let%expect_test "%diagn_show ignores type annots" =
     |}]
 
 let%expect_test "%diagn_show ignores non-empty bindings" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%diagn_show bar { first : int; second : int } : int =
     let { first : int = a; second : int = b } = { first; second = second + 3 } in
     let y : int = a + 1 in
@@ -3250,7 +3250,7 @@ let%expect_test "%diagn_show ignores non-empty bindings" =
     |}]
 
 let%expect_test "%diagn_show no logs" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%diagn_show bar { first : int; second : int } : int =
     let { first : int = a; second : int = b } = { first; second = second + 3 } in
     let y : int = a + 1 in
@@ -3269,7 +3269,7 @@ let%expect_test "%diagn_show no logs" =
     |}]
 
 let%expect_test "%debug_show log level compile time" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%debug3_show () =
     [%log_level
       2;
@@ -3321,7 +3321,7 @@ let%expect_test "%debug_show log level compile time" =
 
 let%expect_test "%debug_show log level runtime" =
   let module Debug_runtime =
-    (val Minidebug_runtime.debug ~values_first_mode:true ~log_level:2 ())
+    (val Minidebug_runtime.debug ~log_level:2 ())
   in
   let%debug3_show () =
     let foo { first : int; second : int } : int =
@@ -3368,7 +3368,7 @@ let%expect_test "%debug_show log level runtime" =
     |}]
 
 let%expect_test "%debug_show PrintBox snapshot" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%debug_show rec loop_highlight (x : int) : int =
     let z : int = (x - 1) / 2 in
     if z = 3 || x = 3 then Debug_runtime.snapshot ();
@@ -3453,7 +3453,7 @@ let%expect_test "%debug_show PrintBox snapshot" =
 
 let%expect_test "%track_show don't show unannotated non-function bindings" =
   let module Debug_runtime =
-    (val Minidebug_runtime.debug ~values_first_mode:true ~log_level:3 ())
+    (val Minidebug_runtime.debug ~log_level:3 ())
   in
   let result =
     [%track_show
@@ -3468,7 +3468,7 @@ let%expect_test "%track_show don't show unannotated non-function bindings" =
         BEGIN DEBUG SESSION |}]
 
 let%expect_test "%log_printbox" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%debug_show foo () : unit =
     [%log_printbox
       PrintBox.init_grid ~line:5 ~col:5 (fun ~line ~col ->
@@ -3599,7 +3599,7 @@ let%expect_test "%log_printbox flushing" =
     |}]
 
 let%expect_test "%log_entry" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%diagn_show _logging_logic : unit =
     let rec loop logs =
       match logs with
@@ -3653,7 +3653,7 @@ let%expect_test "%log_entry" =
 
 let%expect_test "flame graph" =
   let module Debug_runtime =
-    (val Minidebug_runtime.debug_file ~hyperlink:"../" ~toc_specific_hyperlink:"./"
+    (val Minidebug_runtime.debug_file ~values_first_mode:false ~hyperlink:"../" ~toc_specific_hyperlink:"./"
            ~toc_flame_graph:true
            ~backend:(`Html PrintBox_html.Config.(tree_summary true default))
            "test_expect_test_flame_graph")
@@ -3757,7 +3757,7 @@ let%expect_test "flame graph" =
 
 let%expect_test "flame graph reduced ToC" =
   let module Debug_runtime =
-    (val Minidebug_runtime.debug_file ~hyperlink:"../" ~toc_specific_hyperlink:"./"
+    (val Minidebug_runtime.debug_file ~values_first_mode:false ~hyperlink:"../" ~toc_specific_hyperlink:"./"
            ~toc_flame_graph:true ~toc_entry:(Minidebug_runtime.Minimal_depth 1)
            ~backend:(`Html PrintBox_html.Config.(tree_summary true default))
            "test_expect_test_flame_graph")
@@ -3849,7 +3849,7 @@ let%expect_test "flame graph reduced ToC" =
 
 let%expect_test "%debug_show skip module bindings" =
   let optional v thunk = match v with Some v -> v | None -> thunk () in
-  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:true ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
   let%track_sexp bar ?(rt : (module Minidebug_runtime.Debug_runtime) option) (x : int) :
       int =
     let y : int = x + 1 in
@@ -3959,7 +3959,7 @@ let%expect_test "%track_rt_show expression runtime passing" =
     [%log_block
       "test C";
       [%log "line C"]]]
-    Minidebug_runtime.(forget_printbox @@ debug ~global_prefix:"t3" ~log_level:0 ());
+    Minidebug_runtime.(forget_printbox @@ debug ~values_first_mode:false ~global_prefix:"t3" ~log_level:0 ());
   [%expect
     {|
     BEGIN DEBUG SESSION t1
@@ -3978,7 +3978,7 @@ let%expect_test "%debug_show tuples values_first_mode highlighted" =
   let module Debug_runtime =
     (val Minidebug_runtime.debug
            ~highlight_terms:Re.(alt [ str "339"; str "8" ])
-           ~values_first_mode:true ())
+           ())
   in
   let%debug_show bar ((first : int), (second : int)) : int =
     let y : int = first + 1 in
@@ -4069,35 +4069,35 @@ let%expect_test "%logN_block runtime log levels" =
   @@ Int.to_string
        (result
           Minidebug_runtime.(
-            forget_printbox @@ debug ~global_prefix:"for=2,with=default" ())
+            forget_printbox @@ debug ~values_first_mode:false ~global_prefix:"for=2,with=default" ())
           ~for_log_level:2);
   print_endline
   @@ Int.to_string
        (result
           Minidebug_runtime.(
             forget_printbox
-            @@ debug ~values_first_mode:true ~log_level:0 ~global_prefix:"for=1,with=0" ())
+            @@ debug ~log_level:0 ~global_prefix:"for=1,with=0" ())
           ~for_log_level:1);
   print_endline
   @@ Int.to_string
        (result
           Minidebug_runtime.(
             forget_printbox
-            @@ debug ~values_first_mode:true ~log_level:1 ~global_prefix:"for=2,with=1" ())
+            @@ debug ~log_level:1 ~global_prefix:"for=2,with=1" ())
           ~for_log_level:2);
   print_endline
   @@ Int.to_string
        (result
           Minidebug_runtime.(
             forget_printbox
-            @@ debug ~values_first_mode:true ~log_level:2 ~global_prefix:"for=1,with=2" ())
+            @@ debug ~log_level:2 ~global_prefix:"for=1,with=2" ())
           ~for_log_level:1);
   print_endline
   @@ Int.to_string
        (result
           Minidebug_runtime.(
             forget_printbox
-            @@ debug ~values_first_mode:true ~log_level:3 ~global_prefix:"for=3,with=3" ())
+            @@ debug ~log_level:3 ~global_prefix:"for=3,with=3" ())
           ~for_log_level:3);
   (* Unlike with other constructs, INFO should not be printed in "for=4,with=3", because
      log_block filters out the whole body by the log level. *)
@@ -4106,7 +4106,7 @@ let%expect_test "%logN_block runtime log levels" =
        (result
           Minidebug_runtime.(
             forget_printbox
-            @@ debug ~values_first_mode:true ~log_level:3 ~global_prefix:"for=4,with=3" ())
+            @@ debug ~log_level:3 ~global_prefix:"for=4,with=3" ())
           ~for_log_level:4);
   [%expect
     {|
@@ -4321,7 +4321,7 @@ let%expect_test "%logN_block runtime log levels" =
     |}]
 
 let%expect_test "%log_block compile-time nothing" =
-  let module Debug_runtime = (val Minidebug_runtime.debug ()) in
+  let module Debug_runtime = (val Minidebug_runtime.debug ~values_first_mode:false ()) in
   let%diagn_show _logging_logic : unit =
     [%log_level
       0;